+++ src/hotel/model/Individual.java
@@ -2,7 +2,7 @@ package hotel.model;
 
 public class Individual{
 
-    private long id;
+    private int id;
     private String name;
     private String sex;
     private String phoneNumber;
@@ -10,18 +10,18 @@ public class Individual{
     public Individual(){
     }
 
-    public Individual(long id, String name, String sex, String phoneNumber){
+    public Individual(int id, String name, String sex, String phoneNumber){
 	this.id = id;
 	this.name = name;
 	this.sex = sex;
 	this.phoneNumber = phoneNumber;
     }
 
-    public long getId(){
+    public int getId(){
         return id;
     }
 
-    public void setId(long id){
+    public void setId(int id){
         this.id = id;
     }
